+++ boards/PDM/Inc/App/states/App_AllStates.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "App_SharedStateMachine.h"
+
+void AllStatesRunOnEntry(struct StateMachine *state_machine);
+void AllStatesRunOnTick(struct StateMachine *state_machine);
+void AllStatesRunOnExit(struct StateMachine *state_machine);
